chore: app refactor
